+++ udata_front/theme/gouvfr/assets/js/components/MarkdownEditor/ImageModal/ImageModalButton.vue
@@ -16,7 +16,7 @@
 <script setup lang="ts">
 import ImageModal from './ImageModal.vue';
 import TooltipButton from "../EditorButton.vue";
-import { getRandomId } from '@datagouv/components';
+import { getRandomId } from '@datagouv/components/ts';
 import { useI18n } from 'vue-i18n';
 import type { ImageModalForm } from './ImageModal.vue';