+++ udata_front/theme/gouvfr/assets/js/helpers.ts
@@ -1,10 +1,5 @@
-// @ts-ignore
-import type { ResourceType } from "@datagouv/components";
 import RemoveMarkdown from "remove-markdown";
 import { readonly } from "vue";
-import { useI18n } from "vue-i18n";
-import type { ResourceType } from "./types";
-import { lang } from "./config";
 
 export const truncate = (val: string, length = 300) => {
   if (typeof val !== "string") return;
@@ -28,25 +23,6 @@ export const summarize = (value: number) => {
 };
 
 export const CLOSED_FORMATS = readonly(['pdf', 'doc', 'docx', 'word', 'xls', 'excel', 'xlsx'] as const);
-
-export const getResourceLabel = (type: ResourceType) => {
-  const { t } = useI18n();
-  switch(type) {
-    case "main":
-      return t("Main file");
-    case "documentation":
-      return t("Documentation");
-    case "update":
-      return t("Update");
-    case "api":
-      return t("API");
-    case "code":
-      return t("Source code");
-    case "other":
-      return t("Other");
-  }
-}
-
 
 const includeInSubtype = <T, U extends T>(array: ReadonlyArray<U>, value: T): value is U => {
   return array.includes(value as U);
@@ -73,7 +49,7 @@ export const communityResourceUrlRegExp = new RegExp(
 
 export const getResourceIdFromHash = (hash: string, isCommunityResource: boolean): string | null => {
   for (const regex of isCommunityResource ? [communityResourceUrlRegExp] : [resourceUrlRegExp, previousResourceUrlRegExp]) {
-    let [a, foundId, b] = regex.exec(hash) || [];
+    let [_a, foundId, _b] = regex.exec(hash) || [];
     if(foundId) {
       return foundId;
     }